[Rust] Split the glean-core crate into the user-facing part and internal implementation
