+++ glean-core/src/glean_metrics.rs
@@ -1,0 +1,75 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// ** IMPORTANT **
+//
+// This file is required in order to include the ones generated by
+// 'glean-parser' from the SDK registry files.
+
+pub mod pings {
+    use crate::metrics::PingType;
+    use once_cell::sync::Lazy;
+
+    #[allow(non_upper_case_globals)]
+    pub static baseline: Lazy<PingType> = Lazy::new(|| {
+        PingType::new(
+            "baseline",
+            true,
+            true,
+            vec![
+                "background".to_string(),
+                "dirty_startup".to_string(),
+                "foreground".to_string(),
+            ],
+        )
+    });
+
+    #[allow(non_upper_case_globals)]
+    pub static metrics: Lazy<PingType> = Lazy::new(|| {
+        PingType::new(
+            "metrics",
+            true,
+            false,
+            vec![
+                "overdue".to_string(),
+                "reschedule".to_string(),
+                "today".to_string(),
+                "tomorrow".to_string(),
+                "upgrade".to_string(),
+            ],
+        )
+    });
+
+    #[allow(non_upper_case_globals)]
+    pub static events: Lazy<PingType> = Lazy::new(|| {
+        PingType::new(
+            "metrics",
+            true,
+            false,
+            vec![
+                "background".to_string(),
+                "max_capacity".to_string(),
+                "startup".to_string(),
+            ],
+        )
+    });
+}
+
+pub mod error {
+    use crate::metrics::CounterMetric;
+    use crate::{CommonMetricData, Lifetime};
+    use once_cell::sync::Lazy;
+
+    #[allow(non_upper_case_globals)]
+    pub static preinit_tasks_overflow: Lazy<CounterMetric> = Lazy::new(|| {
+        CounterMetric::new(CommonMetricData {
+            name: "glean.error".into(),
+            category: "preinit_tasks_overflow".into(),
+            send_in_pings: vec!["metrics".into()],
+            lifetime: Lifetime::Ping,
+            disabled: false,
+            ..Default::default()
+        })
+    });
+}
